# dist/swagger/swagger.json
{
    "components": {
...
        "schemas": {
            "IApiRoute": {
                "properties": {
                    "group": {
                        "type": "string"
                    },
                    "method": {
                        "type": "string"
                    },
                    "route": {
                        "type": "string"
                    },
                    "scoped": {
                        "type": "string"
                    },
                    "tag": {
                        "type": "string"
                    }
                },
                "required": [
                    "method",
                    "route"
                ],
                "type": "object",
                "additionalProperties": {
                    "type": "string"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "BOS config admin API routes",
        "version": "1.0.0",
        "description": "Welcome to the reference documentation for Portofolio asset manager (PAM) REST API! </br></br>To learn about the common use cases and concept this REST APIs, check out our resource center  (<a>https://resourcecenter.fr.spinalcom.com/ressources-développeur</a>).</br></br>In addition to the API Reference, we also provide complete documentation for adminnistration API, that need to install and implement Spinalcore Digital Twin Operating System on their built environment.",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "Spinalcom",
            "email": "developers@spinalcom.com"
        }
    },
    "openapi": "3.0.0",
    "paths": {
        "/api/v1/pam/create_portofolio_api_route": {
            "post": {
                "operationId": "CreatePortofolioApiRoute",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IApiRoute"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Apis"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IApiRoute"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pam/update_portofolio_api_route/{id}": {
            "put": {
                "operationId": "UpdatePortofolioApiRoute",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IApiRoute"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Apis"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IApiRoute"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pam/get_portofolio_api_route/{id}": {
            "get": {
                "operationId": "GetPortofolioApiRouteById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IApiRoute"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Apis"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/v1/pam/get_all_portofolio_api_route": {
            "get": {
                "operationId": "GetAllPortofolioApiRoute",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "items": {
                                                "$ref": "#/components/schemas/IApiRoute"
                                            },
                                            "type": "array"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Apis"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": []
            }
        },
        "/api/v1/pam/delete_portofolio_api_route/{id}": {
            "delete": {
                "operationId": "DeletePortofolioApiRoute",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Apis"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/v1/pam/upload_portofolio_apis_routes": {
            "post": {
                "operationId": "UploadPortofolioSwaggerFile",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "items": {
                                                "$ref": "#/components/schemas/IApiRoute"
                                            },
                                            "type": "array"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Apis"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pam/create_bos_api_route": {
            "post": {
                "operationId": "CreateBosApiRoute",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IApiRoute"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Apis"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IApiRoute"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pam/update_bos_api_route/{id}": {
            "put": {
                "operationId": "UpdateBosApiRoute",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IApiRoute"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Apis"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IApiRoute"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pam/get_bos_api_route/{id}": {
            "get": {
                "operationId": "GetBosApiRouteById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IApiRoute"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Apis"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/v1/pam/get_all_bos_api_route": {
            "get": {
                "operationId": "GetAllBosApiRoute",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "items": {
                                                "$ref": "#/components/schemas/IApiRoute"
                                            },
                                            "type": "array"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Apis"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": []
            }
        },
        "/api/v1/pam/delete_bos_api_route/{id}": {
            "delete": {
                "operationId": "DeleteBosApiRoute",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Apis"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}